Revert "Feat: Copy several files to destination directory (optionally)"
This reverts commit 855e1150495ea755c2c752ca972d0420c4aa2a2b.
