+++ Editor/IO/Files/CopyFileToDirectory.cs.meta
